+++ crates/astria-sequencer/src/service/info/mod.rs
@@ -221,7 +221,7 @@ mod tests {
     #[tokio::test]
     async fn handle_balance_query() {
         use astria_core::{
-            generated::protocol::accounts::v1 as raw,
+            generated::astria::protocol::accounts::v1 as raw,
             protocol::account::v1::AssetBalance,
         };
 
@@ -287,7 +287,7 @@ mod tests {
 
     #[tokio::test]
     async fn handle_denom_query() {
-        use astria_core::generated::protocol::asset::v1 as raw;
+        use astria_core::generated::astria::protocol::asset::v1 as raw;
 
         let storage = Storage::new_temp().await;
         let mut state_delta = storage.new_delta_of_latest_snapshot();
@@ -331,7 +331,7 @@ mod tests {
 
     #[tokio::test]
     async fn handle_allowed_fee_assets_query() {
-        use astria_core::generated::protocol::asset::v1 as raw;
+        use astria_core::generated::astria::protocol::asset::v1 as raw;
 
         let storage = Storage::new_temp().await;
         let mut state_delta = storage.new_delta_of_latest_snapshot();